requirements.txt
selenium
pytest-html
allure-pytest
pytest-order
pytest-soft-assertions
datetime